requirements.txt
install==1.3.5
python-frontmatter==1.1.0
PyYAML==6.0.1
